+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed resource definition delete setting DELETE flag when resources are still in use
 - Fixed retry deletion of already deleting resource (no longer fails if it is the second to last diskful resource)
+- Satellite now properly merges remote nodes + their data (props, netIfs)
 
 ## [1.33.1] - 2025-12-22